+++ src/main/java/io/github/dsheirer/dsp/filter/BooleanAveragingFilter.java
@@ -37,7 +37,7 @@ public class BooleanAveragingFilter implements Listener<Boolean>
 		mBufferLength = length;
 		
 		/* Round up */
-		mThreshold = (int)( length / 2) + ( length % 2 );
+		mThreshold = ( length / 2) + ( length % 2 );
 	}
 	
 	public void receive( Boolean value )
